Removes folder of unneeded items. Adds Classfai registration to a separate menu
